backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-dotenv==1.0.1
openai==1.58.1
yt-dlp==2024.12.23
aiofiles==24.1.0
